aip_identity.py
#!/usr/bin/env python3
from nacl.signing import SigningKey
import hashlib
import numpy as np

def quantum_model_checksum(spinor: np.ndarray) -> str:
    raw = spinor.flatten().tobytes()
    return hashlib.sha3_512(raw).hexdigest().upper()[:128]

def derive_quantum_identity(spinor: np.ndarray) -> dict:
    seed_hex = quantum_model_checksum(spinor)
    sk = SigningKey(seed_hex.encode()[:32])
    vk = sk.verify_key
    return {
        "seed_hex": seed_hex,
        "ed25519": {
            "private_hex": sk.encode().hex(),
            "public_hex": vk.encode().hex()
        }
    }
