requirements.txt
torch
numpy
pynacl
cryptography
base58
matplotlib
